python-scoring/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
joblib==1.3.2
pandas==2.2.2
numpy==1.26.2
pydantic==2.6.2
python-multipart==0.0.6
scikit-learn==1.3.2
loguru==0.7.0
mysql-connector-python==8.1.1
